Create run_label_app_externally.md
this is an explanation what commands to follow, in order to use your own computer at home or where ever, in order to log into the computer at the wachsbleiche, log into the grid, and run Michael's label app with the pictures loaded from the grid. 

We need this for labeling the images from our computers.